// portal/src/app/shared/models/flights-passangers.ts
export interface FlightsPassengers {
  thingId: string;
  deviceId: string;
  thingName: string;
  thingType: string;
  thingEmployer: string;
  thingPhone: string;
  thingEmail: string;
  thingDepartment: string;
  thingPosition: string;
  thingCompanyName: string;
  thingSourceInfoType: string;
  thingSourceInfoValue: string;
  deviceSourceIdentificator: string;
  latestReportTime: string;
  role: string;
  roleType: string;
  department: string;
  phone: string;
  email: string;
}